fix: prevent tile flicker
This issue describes the precise bug we hit:
https://github.com/godotengine/godot/issues/61220

For now we just don't handle the body_exited signal, and with this
change we never remove an added 'detector' from the simpleTile impl.
We'll need to handle this when we get to character death, but we may end
up with a different method by then anyway. It would be nice to let the
character or tileDetector determine how to interact with the
tilemap (i.e. what color to set the tile to), but it might be odd to
leak that detail.... maybe some tilemap public functions the consumers
can choose between?