feat: TileSetTools tool script
Adds some tiles, draws a quick brick wall and sidewalk.

Impls a tool script for setting collisions on every tile in an atlas by
tile name (atlas name).

Adds a rearranged but not quite working auto-tile bitmasked attempt.

Enables collisions on the Char so that Chars collide with the collision
tiles.